实现了Boar的chase状态，并且完成了patrol和chase的有限状态机的切换。制作了角色血条，精力条，头像部分的ui制作，ui逻辑部分尚未实现

+++ Assets/Scripts/Player/PhysicsCheck.cs
@@ -15,7 +15,7 @@ public class PhysicsCheck : MonoBehaviour
     public LayerMask GroundLayer; //Ö¸¶¨Åö×²²ã¼¶
     public Vector2 bottomOffset;
     public Vector2 leftOffset;
-    public double faceDir;
+    public float faceDir;
     public Vector2 rightOffset; 
     private void Awake()
     {
